Make new module for commissioning package deletion dialog
