+++ www/src/components/editor/CommissioningPackageDeletionDialog.tsx
@@ -1,0 +1,81 @@
+﻿import { Button, Checkbox, Dialog, Table } from "@equinor/eds-core-react";
+import React, { useEffect, useState } from "react";
+import { useCommissioningPackageContext } from "../../hooks/useCommissioningPackageContext.tsx";
+
+interface DeleteDialogProps {
+  isOpen: boolean;
+  onClose: () => void;
+}
+
+const DeleteCommissioningPackageDialog: React.FC<DeleteDialogProps> = ({ isOpen, onClose }) => {
+  const context = useCommissioningPackageContext();
+  const [selectedPackages, setSelectedPackages] = useState<Set<string>>(new Set());
+
+  const handleCheckboxChange = (packageId: string) => {
+    setSelectedPackages((prevSelected) => {
+      const newSelected = new Set(prevSelected);
+      if (newSelected.has(packageId)) {
+        newSelected.delete(packageId);
+      } else {
+        newSelected.add(packageId);
+      }
+      return newSelected;
+    });
+  };
+
+  const handleDelete = () => {
+    selectedPackages.forEach((packageId) => {
+      context?.deleteCommissioningPackage(packageId);
+    });
+    onClose();
+    setSelectedPackages(new Set());
+  };
+
+  useEffect(() => {
+    if (isOpen && context?.activePackage) {
+      setSelectedPackages(new Set([context.activePackage.id]));
+    }
+  }, [isOpen, context?.activePackage]);
+
+  return (
+    <Dialog open={isOpen} onClose={onClose}>
+      <Dialog.Header>
+        <Dialog.Title>Delete Commissioning Packages</Dialog.Title>
+      </Dialog.Header>
+      <Dialog.CustomContent>
+        Choose which commissioning packages to delete:
+        {context?.commissioningPackages.map((commpckg) => (
+          <Table key={commpckg.id}>
+            <Table.Row>
+              <Table.Cell>
+                <div style={{ display: 'flex', alignItems: 'center' }}>
+                  <div
+                    style={{
+                      width: '16px',
+                      height: '16px',
+                      backgroundColor: commpckg.color,
+                      borderRadius: '50%',
+                      marginRight: '8px',
+                    }}
+                  ></div>
+                  <Checkbox
+                    label={commpckg.name}
+                    name="multiple"
+                    checked={selectedPackages.has(commpckg.id)}
+                    onChange={() => handleCheckboxChange(commpckg.id)}
+                  />
+                </div>
+              </Table.Cell>
+            </Table.Row>
+          </Table>
+        ))}
+      </Dialog.CustomContent>
+      <Dialog.Actions>
+        <Button onClick={handleDelete}>Delete</Button>
+        <Button variant="ghost" onClick={onClose}>Cancel</Button>
+      </Dialog.Actions>
+    </Dialog>
+  );
+};
+
+export default DeleteCommissioningPackageDialog;
